[FEAT] Golangci-lint from V1 to V2 (and other clean-up) (#3269)
* refactor: managing linter rules in JSON (v2)

* fix: removed conversions that were not needed

* fix: removed conversions not needed, unwrapped ()

* fix: removed unneeded conversion

* fix: use error built-in checking, added two nolints for future

* fix: removed type conversion, not needed

* fix: added comments, nolint for later

* fix: missing handling the error

* fix: ignore depgaurd for now, can fix later

* fix: use internal errors pkg

* fix: minor fixes, err shadowing, package shadow, etc

* fix: inneffectual

* fix: removed conversions - not needed

* fix: use internal errors package

* fix: added nolint for later testing

* fix: added nolint for later testing

* fix: removed conversion, not needed

* fix: added missing comments

* fix: errors not being handled

* fix: ineffectual assign

* fix: added nolint, needs refactoring

* fix: lots of clean up

* fix: had to disable overflow warnings for now

* fix: added two more nolints for later

* fix: removed deprecated linter

* fix: added nolint for known issue with ripemd160

* fix: out of bounds found

* fix: weak crypto package being used

* fix: target the newest version of golangci-lint

* fix: updated new arg format for v2

* fix: updated cli format

+++ services/blockpersister/utxoset/model/map.go
@@ -75,7 +75,7 @@ func NewSplitSwissMap[K comparableAndHashable, V any](length int) genericMap[K, 
 
 	splitLength := int(math.Ceil(float64(length) / float64(ssm.nrOfBuckets)))
 
-	for i := uint16(0); i <= uint16(ssm.nrOfBuckets); i++ {
+	for i := uint16(0); i <= ssm.nrOfBuckets; i++ {
 		ssm.m[i] = newSwissMap[K, V](splitLength)
 	}
 
@@ -105,7 +105,7 @@ func NewSplitGoMap[K comparableAndHashable, V any](length int) genericMap[K, V] 
 		nrOfBuckets: 1024,
 	}
 
-	for i := uint16(0); i <= uint16(ssm.nrOfBuckets); i++ {
+	for i := uint16(0); i <= ssm.nrOfBuckets; i++ {
 		ssm.m[i] = NewGoMap[K, V]()
 	}
